Added initial functionality for Motif scanning and scoring via PWM class.
Will score simple motif, motif with variable length spacer, or motif with
a defined spacer of variable lengths

+++ src/externDefinitions.cpp
@@ -136,7 +136,7 @@ namespace StochHMM{
     //! \param position  Position in the sequence
     //! \param stateIter Index of state to get weight
     //! \return double value of weight to apply to state at the position
-    double ExDefSequence::getWeight(size_t position, int stateIter){
+    double ExDefSequence::getWeight(size_t position, size_t stateIter){
         if (defs[position]){
             return defs[position]->getWeight(stateIter);
         }
