+++ src/stochMath.cpp
@@ -30,10 +30,10 @@
 namespace StochHMM{
 
     
-    void generateUnknownIndices(std::vector<int>& results, int alphabetSize, int order ,int value){
+    void generateUnknownIndices(std::vector<size_t>& results, int alphabetSize, int order ,int value){
         results.assign(alphabetSize,0);
-        for (int i=0;i<alphabetSize;i++){
-            results[i]= value + i * POWER[order-1][alphabetSize-1];
+        for (size_t i = 0; i < alphabetSize; i++){
+            results[i]= (size_t)value + i * (size_t) POWER[order-1][alphabetSize-1];
         }
         return;
     }
